requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20241021
Django==3.2
django-import-export==3.2.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.3
pytz==2021.1
PyYAML==6.0.1
sqlparse==0.4.1
tablib==3.4.0
typing_extensions==4.7.1
xlrd==2.0.1
xlwt==1.3.0
